Scans using MultiNest

+++ bsfc/bsfc_run_ns.py
@@ -32,12 +32,19 @@ import scipy
 from helpers import bsfc_cmod_shots
 from helpers import bsfc_autocorr
 
-import argparse
+from mpi4py import MPI
+comm = MPI.COMM_WORLD
+rank = comm.Get_rank()
+size = comm.Get_size()
 
+import argparse
 from bsfc_moment_fitter import *
 
+n_hermite=3
+
 # To be removed before public release:
-sys.path.insert(0,'/home/sciortino/usr/pythonmodules/PyMultiNest')
+if '/home/sciortino/usr/pythonmodules/PyMultiNest' not in sys.path:
+    sys.path.insert(0,'/home/sciortino/usr/pythonmodules/PyMultiNest')
 
 parser = argparse.ArgumentParser()
 parser.add_argument("shot", type=int, help="shot number to run analysis on")
@@ -60,36 +67,35 @@ if 'BSFC_ROOT' not in os.environ:
 
 # location of MultiNest chains
 basename = os.path.abspath(os.environ['BSFC_ROOT']+'/mn_chains/c-.' )
+chains_dir = os.path.dirname(basename)
 
 
-# try loading result
-if args.force:
-    # Force us to not load the pickle
-    loaded = False
-    # Remove all the chains
-    import glob
-    fileList = glob.glob(basename+'*')
-    for filePath in fileList:
-        try:
-            os.remove(filePath)
-        except:
-            pass
-
-else:
-    try:
-        with open('../bsfc_fits/mf_NS_%d_tbin%d_chbin_%d.pkl'%(shot,tbin,chbin),'rb') as f:
-            mf=pkl.load(f)
-        loaded = True; print "Loaded previous result"
-    except:
+if rank==0:
+    # try loading result
+    if args.force:
+        # Force us to not load the pickle
         loaded = False
+        # Remove all the chains
+        import glob
+        fileList = glob.glob(basename+'*')
+        for filePath in fileList:
+            try:
+                os.remove(filePath)
+            except:
+                pass
+    else:
+        try:
+            with open('../bsfc_fits/mf_NS_%d_tbin%d_chbin_%d.pkl'%(shot,tbin,chbin),'rb') as f:
+                mf=pkl.load(f)
+            loaded = True; print "Loaded previous result"
+        except:
+            loaded = False
 
-if not loaded:
-    # if this wasn't run before, initialize the moment fitting class
-    mf = MomentFitter(primary_impurity, primary_line, shot, tht=0)
+    if not loaded:
+        # if this wasn't run before, initialize the moment fitting class
+        mf = MomentFitter(primary_impurity, primary_line, shot, tht=0)
 
-    # check that empty directory exists for MultiNest output:
-    chains_dir = os.path.dirname(basename)
-
+    '''
     # delete and re-create directory for MultiNest output
     if os.path.exists(chains_dir):
         if len(os.listdir(chains_dir))==0:
@@ -102,22 +108,32 @@ if not loaded:
     else:
         # if directory does not exist, create it
         os.mkdir(chains_dir)
-
+    '''
+else:
+    mf=None
+    loaded=None
+        
+# broadcast r object to all cores
+mf = comm.bcast(mf, root = 0)
+loaded = comm.bcast(loaded, root = 0)
 
 # ==================================
 
 if loaded==False:
 
     # Do a single spectral fit with nested sampling
-    mf.fitSingleBin(tbin=tbin, chbin=chbin,NS=True, n_hermite=5, n_live_points=400,
-                    sampling_efficiency=0.3, verbose=True)
-
+    mf.fitSingleBin(tbin=tbin, chbin=chbin,NS=True,n_live_points=1000,
+                    sampling_efficiency=0.3,verbose=True,const_eff=True,
+                    n_hermite=n_hermite)
+    
     # save fits for future use
-    with open('../bsfc_fits/mf_NS_%d_tbin%d_chbin_%d.pkl'%(shot,tbin,chbin),'wb') as f:
-        pkl.dump(mf, f)
+    if rank==0:
+        with open('../bsfc_fits/mf_NS_%d_tbin%d_chbin_%d.pkl'%(shot,tbin,chbin),'wb') as f:
+            pkl.dump(mf, f)
 
 if loaded==True:
-
+    # DO NOT try to load and plot with multiple workers (i.e. using mpirun)!
+    
     # load MultiNest output
     mf.fits[tbin][chbin].NS_analysis(basename)
 
@@ -142,17 +158,12 @@ if loaded==True:
     moms = np.average(measurements, 0, weights=sample_weights)
     moms_std = np.sqrt(np.average((measurements-moms)**2, 0, weights=sample_weights))
 
-
     print "Counts = ", moms[0], "+/-", moms_std[0]
     print "v = ", moms[1], "+/-", moms_std[1]
     print "Ti = ", moms[2], "+/-", moms_std[2]
-
-
-# Import mpi4py here to output timing only once
-from mpi4py import MPI
-comm = MPI.COMM_WORLD
-rank = comm.Get_rank()
-size = comm.Get_size()
+    print "ln(ev) = ", mf.fits[tbin][chbin].lnev[0], "+/-", mf.fits[tbin][chbin].lnev[1]
+    print "# Hermite polynomials: ", n_hermite
+    
 
 if rank==0:
     # end time count
